Unify rendering mask data
